+++ src/main/java/roomescape/domain/payment/PaymentRequest.java
@@ -3,7 +3,7 @@ package roomescape.domain.payment;
 public record PaymentRequest(
         String paymentKey,
         String orderId,
-        long amount
+        int amount
 ) {
 
 }
